MOW-1129 Decouple preview cards and material cards
