+++ web/src/main/java/org/geogebra/web/html5/main/PageListControllerInterface.java
@@ -2,7 +2,6 @@ package org.geogebra.web.html5.main;
 
 import org.geogebra.common.kernel.AppState;
 import org.geogebra.common.plugin.EventType;
-import org.geogebra.web.html5.gui.RenameCard;
 
 /**
  * Controller for multipage files
@@ -106,11 +105,4 @@ public interface PageListControllerInterface {
 	 * @return base64 encoded PDF
 	 */
 	String exportPDF();
-
-	/**
-	 * Renaming a slide
-	 * @param card to rename.
-	 * @param title the new title.
-	 */
-	void rename(RenameCard card, String title);
 }
